requirements.txt
scrapy
fake-useragent
requests
django
django-haystack
whoosh
beautifulsoup4
nltk
whitenoise
